header float to int update

+++ src/power_utils.h
@@ -20,7 +20,7 @@ namespace POWER_Utils {
 
     bool    isCharging();
     void    handleChargingLed();
-    double  getBatteryChargeDischargeCurrent();
+    int     getBatteryChargeDischargeCurrent();
     bool    isBatteryConnected();
     void    obtainBatteryInfo();
     void    batteryManager();
